jdbc integration tests
