+++ akka-projection-jdbc/src/it/scala/akka/projection/jdbc/JdbcProjectionSpec.scala
