+++ akka-projection-kafka/src/test/scala/akka/projection/kafka/KafkaSourceProviderSpec.scala
@@ -10,14 +10,12 @@ import scala.concurrent.Await
 import scala.concurrent.Future
 
 import akka.actor.typed.scaladsl.adapter._
-import akka.projection.ContainerTag
 import akka.projection.MergeableOffset
 import akka.projection.kafka.scaladsl.KafkaSourceProvider
 import akka.stream.scaladsl.Source
 import akka.stream.testkit.scaladsl.StreamTestKit.assertAllStagesStopped
 import akka.stream.testkit.scaladsl.TestSink
 
-@ContainerTag
 class KafkaSourceProviderSpec extends KafkaSpecBase() {
   "KafkaSourceProviderSpec" must {
     "resume from provided offsets" in assertAllStagesStopped {
